+++ src/Type/Constants.ts
@@ -4,3 +4,7 @@ export const shortNameLength = 2;
 export const titleTypeMaxLength = 128;
 
 export const colorWarning = '#ffd800';
+
+export const maxRetryAttempts = 10;
+export const retryTimeoutMinMilliseconds = 10;
+export const retryTimeoutVariance = 0.1; // +- 10% variance